feat: add sticky footer and header
