+++ package.json
@@ -23,6 +23,7 @@
     "react-router-dom": "^6.18.0"
   },
   "devDependencies": {
+    "@testing-library/react": "^14.1.0",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@types/react-router-dom": "^5.3.3",
