+++ src/components/Footer/Footer.css
@@ -1,0 +1,21 @@
+body {
+  padding-bottom: 3.75rem;
+}
+#footer {
+  position: fixed;
+  width: 100%;
+  bottom: 0;
+  z-index: 1000;
+  background-color: rgb(224, 224, 224);
+  text-align: center;
+  height: 3.75rem;
+}
+
+@media (min-width: 768px) {
+  body {
+    padding-bottom: 5.125rem;
+  }
+  #footer {
+    height: 5.125rem;
+  }
+}
