+++ src/components/Footer/Footer.tsx
@@ -1,0 +1,13 @@
+import './Footer.css';
+
+const Footer = () => {
+  return (
+    <>
+      <section id="footer">
+        <h1>Footer</h1>
+      </section>
+    </>
+  );
+};
+
+export default Footer;
