+++ src/components/ProductsList/ProductsList.css
@@ -1,0 +1,15 @@
+#contentContainer {
+  position: fixed;
+  top: 3.75rem;
+  bottom: 3.75rem;
+  overflow: scroll;
+  background-color: white;
+  width: 100%;
+}
+
+@media (min-width: 768px) {
+  #contentContainer {
+    top: 5.125rem;
+    bottom: 5.125rem;
+  }
+}
