+++ src/components/ProductsList/ProductsList.tsx
@@ -1,14 +1,15 @@
-import { ChangeEvent, useContext, useEffect, useState } from 'react';
-import ProductItem from './ProductItem';
-import ProductI from '../types/ProductI';
-import fetchProducts from '../utils/fetchProducts';
-import Pagination from './Pagination';
+import { ChangeEvent, useEffect, useState } from 'react';
+import ProductItem from '../ProductItem';
+import ProductI from '../../types/ProductI';
+import fetchProducts from '../../utils/fetchProducts';
+import Pagination from '../Pagination';
 import { NavLink, Outlet, useNavigate, useParams } from 'react-router-dom';
-import SelectPerPage from './SelectPerPage';
-import { SearchContext } from '../context/searchContext';
+import SelectPerPage from '../SelectPerPage';
+import { useSearchContext } from '../../context/searchContext';
+import './ProductsList.css';
 
 const ProductsList = () => {
-  const s = useContext(SearchContext);
+  const { s } = useSearchContext();
   let { page } = useParams();
   const { id } = useParams();
   const [products, setProducts] = useState([] as ProductI[]);
