+++ src/components/Search/Search.css
@@ -1,0 +1,45 @@
+body {
+  padding-top: 3.75rem;
+}
+#searchContainer {
+  top: 0;
+  position: fixed;
+  width: 100%;
+  z-index: 1000;
+  padding: 1.125rem 0.5rem 0 0.5rem;
+  background-color: rgb(224, 224, 224);
+  text-align: center;
+  height: 3.75rem;
+}
+#searchContainer button {
+  background-color: green;
+  color: white;
+  cursor: pointer;
+}
+#searchContainer button:hover {
+  background-color: rgb(0, 101, 0);
+}
+#searchContainer input,
+#searchContainer button {
+  font-size: 1rem;
+  padding: 0.188rem 0.5rem;
+  border: 1px solid green;
+  font-weight: 900;
+}
+
+@media (min-width: 768px) {
+  body {
+    padding-top: 5.125rem;
+  }
+  #searchContainer {
+    padding: 1rem 0.5rem 0 0.5rem;
+    height: 5.125rem;
+  }
+  #searchContainer input,
+  #searchContainer button {
+    font-size: 1.5rem;
+    padding: 0.625rem 1rem;
+    border: 1px solid green;
+    font-weight: 900;
+  }
+}
