+++ src/components/Search/Search.tsx
@@ -1,12 +1,13 @@
 import { FormEvent } from 'react';
-import { useSearchContext } from '../context/searchContext';
+import { useSearchContext } from '../../context/searchContext';
+import './Search.css';
 
 interface SearchProps {
   onSubmitHandler: (event: FormEvent<HTMLFormElement>) => void;
 }
 
 const Search = ({ onSubmitHandler }: SearchProps) => {
-  const s = useSearchContext();
+  const { s } = useSearchContext();
   return (
     <>
       <section id="searchContainer">
